refactor and divide code into modules
