+++ constants.js
@@ -1,0 +1,30 @@
+export const COMMANDS = [
+  'up',
+  'cd',
+  'ls',
+  'cat',
+  'add',
+  'rn',
+  'cp',
+  'mv',
+  'rm',
+  'hash',
+  'compress',
+  'decompress',
+  'os',
+  '.exit',
+];
+
+export const OS_ARGUMEMTS = [
+  '--EOL',
+  '--cpus',
+  '--homedir',
+  '--username',
+  '--architecture',
+];
+
+export const colors = {
+  cyan: '\x1b[36m%s\x1b[0m',
+  yellow: '\x1b[33m%s\x1b[0m',
+  red: '\x1b[31m',
+};
